M3-3868 Change: Include LKE clusters in Support tickets dropdown (#6174)
The only actual change here is that when creating a Support ticket, LKE clusters appear as one of the possible "What is this regarding?" options. However, when I tried to do this I noticed that this select was making a getAll() request on every change. This is old and ridiculous (I wrote it), and the straw that broke the camel's back on this component. I re-wrote it as a function component so that I could use hooks. Then I wrote a bunch of hooks for the major entity types, and combined them into a useEntities hook that provides data and request handlers for each of the types. This way, the support ticket drawer only requests data if it doesn't already have it available in Redux.

Rewrote the tests to use RTL
Adjusted types

* Add useLinodes

* Add useVolumes

* Add useImages

* Add useDomains

* Add useNodeBalancers

* Add useKubernetesClusters

* Add useEntities

* Fix types

* Cleanup

* Fixes 2

* Fixes 3

* Add error handling

* Add union type for EntityType

* Don't get too fancy

* Fix tests and add handling for prefilled ticket data

* Review feedback 1

- Fix copy-paste typos in useKubernetesClusters
- Clear error state in kubernetes.reducer get.start()

* Add better documentation