M3-3996 Re-add token refresh logic
Andrew Sauber added logic in 2018 to refresh a session through Login
before it expired. This was done through an iframe which loaded the
app's null route.

A later update removed this logic, regressing to a situation where
a user can be e.g. typing a support ticket only to have the page suddenly
refresh and their work lost.

The consensus seems to be that the ideal long-term fix for this involves adding
support for PXCE in Login, which would enable us to remove this patch. In the
meantime, however, multiple customers have complained about this, and it is
a genuinely bad experience, so re-adding the patch is worth the lift.