Adjust header padding in Linode Group By Tag
Previously right/left padding was set to 0 for all groups except the first. I think this was probably to make the "No tags" group stand out, since the "No tags" group used to be first in the table.

I tried applying the 0 padding to the *last* group (where the "No tags" group currently lives) but this looked a bit off to me.

Instead I applied the 0 padding everywhere, except for smaller viewports where the cards are used (this looked better to me).