[POC] Speed up the Dashboard  (#6148)
* WIP Reducing Dashboard requests (Volumes)

* Add volumes request to volumeslanding

* Fix logic for updates and store stale data limit in src/constants

* Fix store.helpers test

* Fix imports after merge

* Add volumes to search bar's dependency list

* Update search results when requesting entities

When a request for entities is completed, re-run the search function
in a useEffect so that the results will automatically populate for the
current query.

* Review feedback

- Add .catch() blocks to prevent unhandled error messages
- Refactor LinodeRescue to update Volumes after load

* Fix catch blocks

* Bump react-router-dom to latest patch

* Fix infinite mounting bug

This was a fun one. maybeRenderLoading was wrapping all of LinodesDetail,
and rendering a loading spinner if anything (including Volumes) was loading.
When I added a getAllVolumes() call to componentDidMount() in LinodeRescue,
this created an infinite loop: getAllVolumes() would move volumes to a
loading state, which would unmount LinodeRescue along with the rest of
LinodesDetail to render the spinner. When the request completed, Volumes
were no longer loading, so the spinner was unmounted and LinodesDetail
re-mounted, so the cDM in LinodeRescue fired again, restarting the cycle.

To fix this, I took some of the functional magic out of maybeRenderLoading.
Instead of rendering the spinner itself, it just passes its loading
state as a prop down to LinodeDetail. LD then renders the spinner itself,
allowing us to safely use useReduxLoad at the LD root level.

* Add special handling for small accounts

When getting a single page of Volumes with a page_size of 25 is
enough to retrieve all Volumes data (which is true for most customers,
anyone with 25 or fewer volumes), we update lastUpdated so that
subsequent "do we have this data" requests realize that there's
no reason to make a request.

* Source results from paginated request

Refactored volumes.reducer to use MappedEntityState2, our planned
type for all Redux state. This allows us to reference results
from the Dashboard instead of relying on data.length, which
is inaccurate now that we're only requesting a single page of data.

To support this, added a new version of store.helpers, store.helpers.tmp,
which contains modified versions of the functions in store.helpers, updated
to handle the new data shape.

* Move getVolumesPage into volumeRequests.container

Also refactored the .tmp version of createRequestThunk so that
it returns result.data so that consumers get an Array of what they
want. Had trouble with the typing, hopefully we can update that
once we have a consistent pattern.

* Fixes

- Pass request params to getVolumesPage correctly
- Update volume.reducer so getVolumesPage.done uses result.results

* Review feedback

- Clarify and remove a few comments
- Remove key logic in LinodeDetailNavigation (it was from a
previous attempt to resolve the problem and wasn't actually doing anything)
- Remove unnecessary compose()

* Fix typings in CreateRequestThunk and volumes actions