Fix bug where app would crash if given broken URL query params
When our Redux store is initialized, it looks at the query params
to determine the type of Linode Create flow being specified. This
isn't an ideal pattern, as we read the params even if we're not
entering the app from the create flow.

But the immediate problem is that if you provide broken search params,
such as /search?query=is%linode, the parsing fails and the app crashes.
I added a try/catch block around this logic in the reducer as a patch.