Multiple IP Input errors (#6204)
* Support errors in MultipleIPInput

For Firewall Rules, we need a way to display errors for individual IPs. The MultipleIPInput component dealt with an array of strings, which didn't lend itself well to also keeping track of errors.

I tried a few different things, including a separate `ipErrors` array (and map but managing state separately led to numerous headaches.

The best solution I found was to create a new "ExtendedIP" interface, which has an address and optional `error` string. This way an error is explicitly bound to the IP address it references.

This will come into play when the Firewall Rule Drawer makes use of this new capability.

* Review feedback

Remove unnecessary error prop, since TextField checks !!errorText.