M3-3558 Remove page break on invoices (#6215)
* Add invoiceItem factory and fix type

`unit_price` was incorrectly typed as "null | number" when it should be typed as "null | string". This is also incorrect in the documentation; following up with that team as well.

* Remove extra page for invoice totals table

Previously we were always putting the invoice totals table on its own page. This was unnecessary, as the library seems to handle the page break just fine. I.e. an additional page is added only if needed.

* Adjust padding and widths of invoice totals table

The padding for the first cell was a bit off here, causing the right side to be chopped off. I also adjusted the width of the second cell to tighten the spacing.

* Adjust right-side padding

* aligning left header to match spacing on the right

* Use pageBreak: avoid option

@WilkinsKa1 noticed that the invoices total table was sometimes being split between pages if the invoice had just the right number and type of invoice line items.

This adds the "pageBreak: avoid" which supposedly prevents page breaks unless the actual table is larger than the size of a page (which in our case, it's not).

Co-authored-by: Kayla Wilkins <kwilkins@linode.com>