Fix Volumes creation bug
In an attempt to make the typing work out for the updated createRequestThunk,
I returned response.data from the Thunk's "done" logic. This broke single-entity
requests, such as createVolume, which returns a single Volume object and has
no data property.

I undid this, and tried a different way to resolve the typing. This involved
matching the GetAllData type with ResponsePage, which assumes an array
of its type argument (ResponsePage<Linode> returns Linode[] in its data field).

For now, manually did then(result => result.data) in the Volumes and NBs
handling in useEntities so that they return an array as intended.