Insert tax ID below headers
Previously the "addTitles" method, used to insert the Tax ID into the document, used a constant Y value.

This meant that if a user had a long address, there would be a collision.

To fix this, the "addLeftHeader" and "addRightHeader" methods now return the Y value of the last line they inserted. We then pass the max of these (either left OR right header Y coordinate) to the "addTitles" method, which uses this dynamic value instead of the constant.