Allow custom empty messaging on EventsLanding
Allow the consumer to specify a message to display in the empty state. This means the Activity tab of Linode Details can say "No recent activity for this Linode." instead of "No events on your account".

Also take into account the number of events when determining whether to display "No more events to show." If there are no events, the message above suffices.