Fix skeleton loader on EventsLanding (#6245)
Previously the skeleton loaders on the EventsLanding page were questionable at best.

The EventsLanding table has 4 columns, but one of them is used for the entity icon, so to the user it appears like there's only 3 columns.

We still want the TableRowLoading component to span 4 columns, but we only want to show 3 skeleton loaders. Previously, `colSpan` was overloaded in the sense that it handled both of these things. Now, `colSpan` can be used as intended and `numberOfColumns` can be used to specify the number of skeleton loaders.

If `numberOfColumns` is NOT provided but `colSpan` IS provided, we use `colSpan` for both purposes, like it was previously.