Fix regression in Domains drawers
DomainDrawer was using pathOr for its custom connect logic, so
it didn't pick up on the changes to the Domains redux state. This
led to a regression where domain data in the drawer was always defaulted
to undefined. This PR uses nullish coalescing to fix this and prevent
similar issues in the future.