Remove UNSAFE_componentWillReceiveProps in DomainRecordsDrawer
This drawer was using cWRP to set its field state; afaict, since
records are updated with a live request when landing on DomainDetail,
and are not stored in Redux, this setup was made to ensure that
late-loaded data would be available in the drawer.

Not only was it not ideal to be using a deprecated lifecycle method,
but the refactors in this PR broke the logic, (possibly) because
Redux state updates were triggering the cWRP on submit. Whatever the
cause, after pressing the submit button in the drawer, the fields would
reset to their values from props while the button entered a loading state,
so it looked to the user as if the changes had been lost.

I replaced this with componentDidUpdate, and used our standard
"is the drawer opening" logic to set off the fields state update.