Add timeout request to resolve race(?) condition
When restoring from a backup, there seems to be a race condition:
we receive a percent_complete: 100 event, make a final request,
but the Linode still reports its status as "restoring". This is true
for >5 seconds after the final event rolls in. I added a final request
delayed 10 seconds, which is not pretty but it resolves the issue.

It's still unclear why the final `status: finished` event isnt' registering
with our system, even though we continue to poll.