Store support ticket input in local storage
If a user's session expires while working on a support ticket,
or they accidentally close the tab, their work will be lost.

The first situation can and should be solved by some kind of token
refresh logic, but in the meantime and for the second situation,
I added logic to store the values in local storage.

- Added supportText field to utilities/storage.ts
- Added logic in SupportTicketDrawer to read and write values

Local storage will be cleared:
 - when the drawer is closed (including after a successful submit)
 - when the user explicitly logs out