Differentiate between resetting and clearing the drawer state
Added a prop to resetDrawer and resetTicket helpers. If true, it will
clear all the drawer values, which in turn will cause local storage to be
cleared. This case is used when closing the drawer (including after submitting
a ticket). If false (the default case), these methods will do what they've always
done, which is reset errors and attachments and reset the ticket content to
values from props (or if those aren't provided, then local storage values).