Fix Domain root-level loading logic
We had an infinite request loop caused by the loading spinner
that I put above the Route components. To fix this, I moved the
logic into the details route. I added a case to the "not found"
logic to avoid flashing a 404 message when we hadn't yet completed
a domains request.

Also removed a request for domains on the detail page, since we're now
making one already from the root level.