Use DisplayPrice
I didn't use DisplayPrice originally because the font size was
fixed and I needed something smaller. But on further consideration
it's more important to be consistent in displaying currency,
especially within the same view. So I added a prop to DisplayPrice
to override the font size and used that prop in NodePoolSummary.