Fix Longview NGINX bug
A user reported that their NGINX tab in Longview was crashing.
This is due to a badly reformed response from the Longview API.
Classic Manager's Longview client handled for this case but did
not display data for the malformed request (probably a try/catch somewhere).

It turns out the crash was because we weren't manually type checking
the arguments to appendStats() in longview/shared/utilities. The bad
request was sending an array rather than an object to that method
and Object.keys().reduce() was crashing. Added an isArray() check
to prevent this.

We actually display data in this case, though it can't possibly be
entirely accurate since the "bad" data is being ignored. I'm not sure
whether the best thing to do is check the shape higher up and blank
out the graphs if the shape doesn't match (basically what Classic does)
or leave it as it is now, where graphs are displayed but some data
may be missing.