[Performance] Update Domains requests (#6216)
* refactor Domains actions/reducer/requests for addPage and MappedEntityState2

* Bug fixes in Domains reducer and add unit tests

* Update containers

Add getAllDomains and getDomainsPage to the Domains container
and use them appropriately on the Dashboard, DomainLanding, and DomainDetail

* Fix usage of Domains entities entry in SupportTicketsDrawer

* Fix usage of Domains entities entry in SupportTicketsDrawer

* Remove domains logic from AuthenticationWrapper and DataLoadedListener

* Review feedback

- Clean up typos
- Add error handling/clearing on .started() and .failed() actions for getPage
- Adjust onError funtion in store.helpers.tmp so that it preserves existing
errors in other fields (so updating error.read doesn't clear error.create)

* Fix tests

Move error clearing from getPage.done to getPage.start to match
existing patterns.

* Fix regression in Domains drawers

DomainDrawer was using pathOr for its custom connect logic, so
it didn't pick up on the changes to the Domains redux state. This
led to a regression where domain data in the drawer was always defaulted
to undefined. This PR uses nullish coalescing to fix this and prevent
similar issues in the future.

* Add missing request for domains to domainsdetail/index.tsx

* Remove UNSAFE_componentWillReceiveProps in DomainRecordsDrawer

This drawer was using cWRP to set its field state; afaict, since
records are updated with a live request when landing on DomainDetail,
and are not stored in Redux, this setup was made to ensure that
late-loaded data would be available in the drawer.

Not only was it not ideal to be using a deprecated lifecycle method,
but the refactors in this PR broke the logic, (possibly) because
Redux state updates were triggering the cWRP on submit. Whatever the
cause, after pressing the submit button in the drawer, the fields would
reset to their values from props while the button entered a loading state,
so it looked to the user as if the changes had been lost.

I replaced this with componentDidUpdate, and used our standard
"is the drawer opening" logic to set off the fields state update.

* remove loading state for domains records request

* Fix Domain root-level loading logic

We had an infinite request loop caused by the loading spinner
that I put above the Route components. To fix this, I moved the
logic into the details route. I added a case to the "not found"
logic to avoid flashing a 404 message when we hadn't yet completed
a domains request.

Also removed a request for domains on the detail page, since we're now
making one already from the root level.