Feature/preserve ticket text (#6253)
* Store support ticket input in local storage

If a user's session expires while working on a support ticket,
or they accidentally close the tab, their work will be lost.

The first situation can and should be solved by some kind of token
refresh logic, but in the meantime and for the second situation,
I added logic to store the values in local storage.

- Added supportText field to utilities/storage.ts
- Added logic in SupportTicketDrawer to read and write values

Local storage will be cleared:
 - when the drawer is closed (including after a successful submit)
 - when the user explicitly logs out

* Adjustments

- Make the setting logic unconditional, to allow the user to clear
their input and have that be stored in localStorage.
- Add a reducer case to authentication.reducer to make sure any stored
input is cleared when the user explicitly logs out. Previously the same
reducer case was used for expireTokens and logOut actions, but I think
that would defeat the purpose (if the user's tokens expire we want to
preserve their text).

* Differentiate between resetting and clearing the drawer state

Added a prop to resetDrawer and resetTicket helpers. If true, it will
clear all the drawer values, which in turn will cause local storage to be
cleared. This case is used when closing the drawer (including after submitting
a ticket). If false (the default case), these methods will do what they've always
done, which is reset errors and attachments and reset the ticket content to
values from props (or if those aren't provided, then local storage values).

* Fix test case

- Added a fallback to the getStorage() method for supportText
- Added === undefined check to the existing === null check in getStorage(),
so if the value is not in local storage it'll also get the default. I don't think
this affects actual usage, but our mock local storage implementation seems to break
on nested keys without this.

* Review feedback

- Use ?? instead of || when checking prefilled description values
from props. This allows other components to manually pass description=""
without being overridden by any possible values in localStorage

- debounced the function that writes to localStorage.