M3-4066 checkout bar 2 (#6265)
* Layout for updated KubeCheckoutBar

* Add handlers to checkout menu

* Add trash icon

* Add warning for small node pools

* Add initial unit tests

* Use DisplayPrice

I didn't use DisplayPrice originally because the font size was
fixed and I needed something smaller. But on further consideration
it's more important to be consistent in displaying currency,
especially within the same view. So I added a prop to DisplayPrice
to override the font size and used that prop in NodePoolSummary.

* Update __data__/types to repair broken linode resize test

* styling updates part1

* Fix merge issues

* Bump total storage of kubeUtils test to match updated Linode types in __data__

Co-authored-by: Kayla Wilkins <kwilkins@linode.com>