Fix handling of local storage on logout (#6276)
I originally thought that handleLogout was called on manual logout, and
refreshTokens was called when a user's tokens had expired. On closer
inspection, I realized that refreshTokens was never called at all.

This meant that the original case the localStorage cache of support
tickets was intended to solve wouldn't actually work, since when the
user is redirected to Login the values are cleared. To resolve this,
I deleted the unused refreshTokens action and instead cleared localStorage
from inside the Logout component, which is the actual logic triggered uniquely
when a user clicks the logout button.