Fix create flow bug and add general error handling
- Added a Notice to display errorMap.none (general error)
- Adjusted the updateFor dependencies in the checkout bar. This
fixes a bug introduced by the conversion to an FC. The functions
are no longer instance methods, so they change to reflect current
component state. Since the checkoutBar wasn't listening for changes
to these functions, when submitted older versions of the functions
were called with old state.

Example:

1. Fill in Region and Version but don't add node pools
2. Submit the form
3. Observe: field errors are displayed for both version and node pools,
even though version is correct.

I considered removing renderGuard from the components in this form,
but that actually slows down the form significantly. I think the ideal
thing is a bunch of React.memo and React.useCallback, but that can wait.