Make sure window.paypal is defined before using it (fixes crash) (#6282)
* Make sure window.paypal is defined

This in theory fixes a unreproducible bug in which the PayPal script is blocked, but react-async-script-loader thinks it was loaded successfully. If that happens, the app will crash because we try to use window.paypal, which is undefined unless the actual PayPal script is loaded correctly.

We've heard a report of this happening, and I believe it had to do with blocked pages at the user's network level. In this case, the PayPal script was routed through a filter, which returned a 200 but without the correct content.

With this fix, if something like the above happens, PayPal will simply not appear as a payment option (instead of the app crashing).

* Use hasOwnProperty instead

[review feedback]