Do not pass theme object in RenderGuard
This fixes an issue where many of our components were being rendered down to divs with a "theme=[object Object]" attribute.

This was because RenderGuard was passing all its props... including the theme object... to each component.

In particular, <Grid /> uses RenderGuard, so this appeared A LOT throughout the app.

In addition to being ugly markup, I'm thinking this may actually  increase performance.