Fix NodeBalancer error handling
We received reports of users with very old NodeBalancers being unable
to edit their configs. Spaces used to be permitted in NB labels,
but have since been restricted in both Classic and the API.

I wasn't able to reproduce the specific issue described, even when
creating a NB with an illegal name in devenv; however, I did find
that our error handling for the editable breadcrumb was incomplete.

I fixed this, added a general error Notice (don't know how that would
happen but just in case), and used the errorMap pattern instead of
the older getErrorFor.