Review feedback
@Jskobos noticed that error messages would appear on the wrong rule if another rule was pending deletion.

This is because we DON'T send rules pending deletion in the PUT request (that's how a DELETE works...).

So the API sets the error index based on the rules we gave it, which will be different than the rules we have in state.

To fix this, I rearranged some of the "prepare rules" logic, added the original index as an attribute on the Extended Rules, and referenced that original index in the .catch logic.