[FW] Client-side form validation (#6194)
* Error parser

* Map errors to rows

* Add errors to drawer

* Cleanup

* Fix rebase issue where handleIPChange was left unused

* Form validation

* Skip tests for now

* Fix tests

* Support errors in MultipleIPInput

For Firewall Rules, we need a way to display errors for individual IPs. The MultipleIPInput component dealt with an array of strings, which didn't lend itself well to also keeping track of errors.

I tried a few different things, including a separate `ipErrors` array (and map but managing state separately led to numerous headaches.

The best solution I found was to create a new "ExtendedIP" interface, which has an address and optional `error` string. This way an error is explicitly bound to the IP address it references.

This will come into play when the Firewall Rule Drawer makes use of this new capability.

* Form validation

* Cleanup and add tests

* Small change

* Review feedback

@Jskobos noticed that error messages would appear on the wrong rule if another rule was pending deletion.

This is because we DON'T send rules pending deletion in the PUT request (that's how a DELETE works...).

So the API sets the error index based on the rules we gave it, which will be different than the rules we have in state.

To fix this, I rearranged some of the "prepare rules" logic, added the original index as an attribute on the Extended Rules, and referenced that original index in the .catch logic.

* Review feedback

Stricter regex for validating ports

* Change variable names for clarity

* Use Ramda omit