M3-4146 Update: Minimum payment placeholder text
We show "5.00 minimum" in the number input for choosing a payment amount.
We do this under all conditions, even though the API will accept a smaller
payment if the balance is greater than 0 but less than 5. Our placeholder
has confused a few customers with small balances, who didn't think
it was possible to only pay their balance.

Paypal (I think) always requires a minimum of $5. For all other payments
we follow the API's validation logic.