Dispatch done action even if no buckets (#6338)
The getAllBucketsForAllClusters `done` action was only dispatched if data.length > 0, which meant that if you didn't have any buckets, you'd see an eternal loading spinner on the BucketsLanding page.

I also realized another issue, which is that if you only have buckets in Newark, and the Newark request fails, you'll see the "Empty" state, without an error message. I abstracted the error display out to its own component and include it in the main content as well as the "Empty" state.