M3-3798 Address Xterm flow control errors
We were receiving Sentry reports of xterm.js reaching its flow
control limit and discarding write data. After talking with Abe,
this is a normal thing and an appropriate way to handle it. Since
the underlying screen session has a log, no data is actually lost.

I've been unable to replicate this effect (Abe suggested cat /dev/urandom
but that doesn't trigger it). He suggested that we do what they've
previously done for Glish, which is reload the Lish window when one
of these errors occurs. This will bring the window back in sync with
the screen session.

I also moved the events polling initialization so that it doesn't happen when the user
is loading Lish. There's no reason to poll events from inside a Lish console.