Hide Recycle Node action menu
Temporarily removing the action menu from the node pool rows in
cluster detail view. The recycle functionality on the backend isn't
working properly, so we will hide it until the backend is prepared.

While I was in there, I deleted a bunch of old files from the original
LKE UI that are no longer used (dead code).