getAllBuckets requests accept cluster IDs
My previous approach to this didn't account for the fact that different environments have different clusters. This meant that non-prod environments were requesting the production clusters.

I'm choosing to view this as a sign that my previous intuition to use the API response at object-storage/clusters as the source of truth. We will incur a performance penalty, since Clusters need to be requested before buckets. In my experience this is a fast endpoint, so it should only add ~100ms to the OBJ bucket landing page.