Fix CVV logic
Asked the API team about this and went over the code with them.
The API never requires a CVV, and has no way of determining if
a payment was rejected because of a CVV. Instead, they submit the CVV
if the user provides it, and if the payment is rejected AND the
CVV was not provided, they return an error message suggesting
"Perhaps you should use a CVV" (not kidding about the perhaps).

In light of this, I removed the CVV error variable. We already show
the API error response when submitting a payment, and that seems
to be the best we can do.

This may change once we update our payment processor, stay tuned.