Update Currency component
We needed to be able to use a custom format for currency display. The library we were using allows you to do this, but is poorly documented. There are other problems too, like the fact that it is out of date, there aren't many tests, etc. Since we weren't using many features of the library, I opted to go with a custom implementation instead.