Fix useEffect bug
When I updated the useEffect in ConfigNodeIPSelect, it introduced a bug
related to the function in the deps array (props.handleChange). This
was a small thing to fix, but I decided that it wasn't ideal to have
a lower-level component handling workflows and state as it was.

The IPSelect component was passed a "workflow" prop which was only
used to reset the node's selected IP address when the region changed, if
the user was in the create workflow. It makes more sense imo to have the
create flow manage this behavior itself rather than tell the IPselect
it's creating a NB and have the child take it from there. I added
a method to reset all IP addresses for all configs/nodes in NodeBalancerCreate,
and called it if the selected region changed.