Fix StackScripts bugs (#6359)
Fix bug when deleting a StackScript:

1. Scroll down past the waypoint and delete one of the stackscripts there.
2. Observe: you are redirected to the top of the page, but are unable
to trigger infinite scroll.

This happened because the delete handler was requesting the first page
of data (correctly) but wasn't updating the current page in state.

Fix error and loading handling when deleting:

There were no submitting states for the confirmation dialogs. Added one
for the delete dialog. In addition, the "fieldError" in state wasn't displayed
anywhere, and in any case it's our pattern to display errors like this inside
the confirmation dialog. Adjusted the dialog state and props to account for this.