M3-4107 [Billing] Activity Feed (#6346)
* WIP activity feed

* Fix component props

* Add error handling

* Clean up and fix test

* Fix props.children in TableContentWrapper

* Wrap test in async act

Rendering the component was resulting in an "Updated outside of act()" warning. Solution from https://github.com/facebook/react/issues/15379#issuecomment-482101020.

* Use wait() instead of act()

[review feedback]

* Update Currency component

We needed to be able to use a custom format for currency display. The library we were using allows you to do this, but is poorly documented. There are other problems too, like the fact that it is out of date, there aren't many tests, etc. Since we weren't using many features of the library, I opted to go with a custom implementation instead.

* Adjustments to currency display

All payments and credits are displayed as negative numbers wrapped in parentheses.

* Remove react-currency-formatter

* Move pagination footer

[review feedback]
Move pagination footer inside <Paper />

* Use Intl.NumberFormatter to handle commas

A unit test caught a regression, which is that quantities >= 1000 were no longer separated by a comma. I used the Intl.NumberFormatter to handle for this, and added a unit test to <Currency/>

* Right-align amount column

* Add tests for BillingActivityPanel