Fix TFA disable bug
A refactor of TFA a year ago (looks like #4298) introduced a bug
that's gone unnoticed until now:

If you start the TFA process but don't confirm it (by scanning the code
and entering the generated token), the toggle should...well, toggle.
Right now, pressing the button in mid-process does nothing.

I believe this restores the original behavior; if the user starts the process,
enableTwoFactor is called, but this only generates a secret, it does not
actually enable anything until the process is complete.