Remove HTML from version string
In d72dcdd1afda418d6fbb53e7b64703ccc564ba89, HTML was added to the version string. I believe this was done so that it could be targeted for styling.

We actually lost the styling in a recent update to the markdown conversion anyway (not allowing classes). This removes the HTML (which was showing up in emails) and targets the last <p> of .formatted-text so apply the styling that was intended.