Remove HTML from version string (#6397)
In https://github.com/linode/manager/pull/5355, HTML was added to the version string. I believe this was done so that it could be targeted for styling.

We actually lost the styling in a recent update to the markdown conversion anyway (by not allowing classes). 

This removes the HTML (which was showing up in emails) and targets the last `<p>` of `.formatted-text` and applies the styling that was intended.