[Billing] Download PDF functionality (#6380)
This adds functionality for downloading Invoice and Payment PDFs.

Also new is the `useSet()`hook, which exposes a simple API for using the Set data structure in Function Components. This is sort of a POC; I'm curious to know your thoughts.

It seems like a good place to try this out, because I needed to keep track of which Invoices/Payments had error/loading states (for PDF downloads). I could have modeled the data such that invoices/payments were extended with `pdfError` and `pdfLoading` attributes, but it seemed like overkill. This lets you do:

```typescript
invoicePDFLoading.add(invoiceId);
```

to register a loading state for an Invoice, for example. And to check it:

```typescript
invoicePDFLoading.has(thisInvoiceId);
```

* Small tweaks

- Rename “Activity” to “Billing & Payment History”
- Remove PDF icon in favor of “View PDF” button
- Negative Payments labeled “Refund to Card” instead of “Credit”

* Start of PDF functionality

* WIP PDF functionality

* Use set hook

* Fix BillingDetail test

* Review feedback

- Use one set for errors and one for loading
- Clear sets when changing view controls
- Also fix bug where the error wasn't being cleared on submit.