[SDK] Add tests (#6396)
* First working SDK test

- Add axios-mock-adapter as a dev dependency
- Wire up a single test for the setURL method

* Add tests for remaining composition helpers

Fixed a sort-of bug in setHeaders(), which was overwriting the full
headers object. This is a problem because setXFilter() sets
headers['X-Filter'], so if setXFilter() is called before setHeaders()
it won't have any effect.

Used some over(lens, merge) magic to fix this. This also necessitated
defaulting headers to {} in the config reducer.

* WIP schema tests

* Complete schema validation tests

* Delete unused support SDK tests

* Exclude compiled files from testing