Insert HTML before displaying version
I overshot a bit when I tried to fix this issue before. I applied the CSS to the last "p" element of a Support Ticket. That's a problem though, since e.g. Ticket Replies do not have the version string appended to the description.

This inserts a "span" with a class for the CSS to select on.