HTML in support tickets [again] (#6410)
My solution to this issue in https://github.com/linode/manager/pull/6397 overshot a bit. There, I applied the CSS to the last "p" element of a Support Ticket. That's a problem though, since not everything has the version string attached (replies, subsequent updates, tickets opened by Linode).

This inserts a "span" with a class for the CSS to select on.

* Insert HTML before displaying version

This inserts a "span" with a class for the CSS to select on.

* Fix test

* Add tests for allowed class