Clarify auto-resize text (#6416)
* Clarify auto-resize text

A customer had expressed that our messaging about auto disk
resizes when resizing a Linode was unclear. We included relevant
information in tooltips (when the new plan was smaller or the user
had custom disks). I added some clarifying comments to the text
next to the checkbox, and left the tooltips as they were.

* Undo a bad thing

* Refactor a bit